pyproject.toml
[project]
name = "2025-11-09-filmeve"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "catboost>=1.2.8",
    "joblib>=1.5.2",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "plotly>=6.5.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "streamlit>=1.51.0",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "xgboost>=3.1.1",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "pytest>=9.0.0",
    "ruff>=0.14.4",
]

[tool.pytest.ini_options]
pythonpath = "src"
